requirements.txt
faster-whisper
soundcard
soundfile
numpy
colorama
openai
python-dotenv